gx10/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
python-dotenv>=1.0.0
